rename: user api 명세서 패키지 구조 변경
