+++ src/main/java/com/map/gaja/user/presentation/api/UserController.java
@@ -2,7 +2,7 @@ package com.map.gaja.user.presentation.api;
 
 import com.map.gaja.user.presentation.dto.request.Req;
 import com.map.gaja.user.presentation.dto.response.Res;
-import com.map.gaja.user.presentation.swagger.UserApiSpecification;
+import com.map.gaja.user.presentation.api.specification.UserApiSpecification;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
